apps/api/users/factories.py
"""
Factory classes for generating test data.

Uses factory_boy to create realistic test data for the users app.
"""

import factory
from django.contrib.auth import get_user_model

User = get_user_model()


class UserFactory(factory.django.DjangoModelFactory):
    """Factory for creating User instances."""

    class Meta:
        model = User

    email = factory.Sequence(lambda n: f"user{n}@example.com")
    username = factory.Sequence(lambda n: f"user{n}")
    password = factory.PostGenerationMethodCall("set_password", "testpass123")
    is_active = False  # Default for new users
    is_staff = False
    is_superuser = False
