apps/api/requirements/dev.txt
# Include base requirements
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
freezegun==1.4.0
factory-boy==3.3.0
faker==20.1.0
schemathesis==3.27.1

# Linting & Code Quality
ruff==0.1.8
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
bandit==1.7.5

# Development tools
ipython==8.18.1
ipdb==0.13.13
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

black==25.9.0
flake8==7.3.0
mypy==1.7.1
isort==7.0.0
